Add files via upload
Función fraccional y presentación

+++ FuncionMultiplicacion.h
@@ -1,8 +1,8 @@
-float multiplicarNumeros(int A,int B){
+int multiplicarNumeros(int A,int B){
     printf("Funcion Multiplicacion: \n");
 
     int multiplicarNumeros= A*B;
-    printf("La multiplicacion de los numeras da como resultado: %f \n",multiplicarNumeros);
+    printf("La multiplicacion de los numeras da como resultado: %d \n",multiplicarNumeros);
 
 
 }
