+++ FuncionResta.h
@@ -1,8 +1,8 @@
-float restarNumeros(int A,int B){
+int restarNumeros(int A,int B){
     printf("Funcion Resta: \n");
 
     int restarNumeros = A-B;
-    printf("La resta de los numeras da como resultado: %f \n",restarNumeros);
+    printf("La resta de los numeras da como resultado: %d \n",restarNumeros);
 
 
 }
